Organize ui package by app's screens.
