+++ app/src/main/java/ru/sandbox/androidacademyapp/ui/moviedetails/ActorsAdapter.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.ui
+package ru.sandbox.androidacademyapp.ui.moviedetails
 
 import android.view.LayoutInflater
 import android.view.View
