+++ app/src/main/java/ru/sandbox/androidacademyapp/ui/moviedetails/ActorsItemDecoration.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.ui
+package ru.sandbox.androidacademyapp.ui.moviedetails
 
 import android.graphics.Rect
 import android.view.View
