+++ app/src/main/java/ru/sandbox/androidacademyapp/ui/movies/MoviesItemDecoration.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.ui
+package ru.sandbox.androidacademyapp.ui.movies
 
 import android.graphics.Rect
 import android.view.View
